+++ src/tests/superValidate.test.ts
@@ -21,20 +21,7 @@ import { zod, zodToJSONSchema } from '$lib/adapters/zod.js';
 import { z } from 'zod';
 
 import { valibot } from '$lib/adapters/valibot.js';
-import {
-	object,
-	string,
-	email,
-	minLength,
-	array,
-	integer,
-	number,
-	minValue,
-	date,
-	optional,
-	regex,
-	transform
-} from 'valibot';
+import * as v from 'valibot';
 
 //import { ajv } from '$lib/adapters/ajv.js';
 //import type { JSONSchema } from '$lib/jsonSchema/index.js';
@@ -204,13 +191,13 @@ describe('Arktype', () => {
 /////////////////////////////////////////////////////////////////////
 
 describe('Valibot', () => {
-	const schema = object({
-		name: optional(string(), 'Unknown'),
-		email: string([email()]),
-		tags: array(string([minLength(2)]), [minLength(3)]),
-		score: number([integer(), minValue(0)]),
-		date: optional(date()),
-		nospace: optional(string([regex(nospacePattern)]))
+	const schema = v.object({
+		name: v.optional(v.string(), 'Unknown'),
+		email: v.string([v.email()]),
+		tags: v.array(v.string([v.minLength(2)]), [v.minLength(3)]),
+		score: v.number([v.integer(), v.minValue(0)]),
+		date: v.optional(v.date()),
+		nospace: v.optional(v.string([v.regex(nospacePattern)]))
 	});
 
 	describe('Introspection', () => {
@@ -219,6 +206,26 @@ describe('Valibot', () => {
 
 	describe('Defaults', () => {
 		schemaTest(valibot(schema, { defaults }), undefined, 'simple');
+	});
+
+	it('should produce a required enum if no default', () => {
+		const schema = v.object({
+			enum: v.picklist(['a', 'b', 'c']),
+			enumDef: v.optional(v.picklist(['a', 'b', 'c']), 'b')
+		});
+
+		const adapter = valibot(schema);
+		expect(adapter.jsonSchema.required).toEqual(['enum']);
+		expect(adapter.defaults).toEqual({
+			enum: 'a',
+			enumDef: 'b'
+		});
+
+		// Change defaults
+		adapter.defaults.enum = '' as 'a';
+
+		const a2 = valibot(schema);
+		expect(a2.defaults.enum).toBe('');
 	});
 });
 
@@ -343,6 +350,30 @@ describe('Zod', () => {
 		num;
 	});
 
+	it('should produce a required enum if no default', () => {
+		enum Fruits {
+			Apple = 7,
+			Banana = 8
+		}
+
+		const schema = z.object({
+			nativeEnumInt: z.nativeEnum(Fruits),
+			nativeEnumString: z.nativeEnum({ GRAY: 'GRAY', GREEN: 'GREEN' }).default('GREEN'),
+			enum: z.enum(['a', 'b', 'c']),
+			// eslint-disable-next-line @typescript-eslint/no-explicit-any
+			enumDef: z.enum(['a', 'b', 'c']).default('' as any)
+		});
+
+		const adapter = zod(schema);
+		expect(adapter.jsonSchema.required).toEqual(['nativeEnumInt', 'enum', 'enumDef']);
+		expect(adapter.defaults).toEqual({
+			nativeEnumInt: Fruits.Apple,
+			nativeEnumString: 'GREEN',
+			enum: 'a',
+			enumDef: ''
+		});
+	});
+
 	schemaTest(zod(schema));
 });
 
@@ -377,8 +408,8 @@ describe('Schema In/Out transformations', () => {
 	});
 
 	it('does not fully work with Valibot', async () => {
-		const schema = object({
-			len: transform(string(), (s) => s.length)
+		const schema = v.object({
+			len: v.transform(v.string(), (s) => s.length)
 		});
 
 		// @ts-expect-error Using schema Out type as In - Not allowed